% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.roxy.R
\name{add.roxy}
\alias{add.roxy}
\title{add.roxy}
\usage{
add.roxy(obj, file = FALSE)
}
\arguments{
\item{obj}{the function to create documentation for}

\item{file}{if TRUE write to file (R/{function name}.R), Default: FALSE}
}
\value{
TRUE if successful, FALSE if not
}
\description{
adds roxygen documentation headers to the top on R function and
optionally write this to a file in the R folder.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

}
\seealso{
\code{\link[sinew]{makeOxygen}}
\code{\link[formatR]{tidy_source}}
}
