% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dataset.R
\name{add_dataset}
\alias{add_dataset}
\title{add_dataset}
\usage{
add_dataset(name)
}
\arguments{
\item{name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

}
\seealso{
\code{\link[glue]{glue}}
\code{\link[icesTAF]{taf.boot.path}}
}
