% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_init.R
\name{git_init}
\alias{git_init}
\title{git_init}
\usage{
git_init(path = ".", commit = TRUE)
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: '.'}

\item{commit}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
no return value
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
if (interactive()) {
  # initialise a git repository and stage all files
  git_init(commit = FALSE)
}
}
\seealso{
\code{\link[git2r]{init}},\code{\link[git2r]{add}},\code{\link[git2r]{c("commit", "commit")}}
}
