% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed_packages.R
\name{installed_packages}
\alias{installed_packages}
\title{installed_packages}
\usage{
installed_packages()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
