% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.dir.R
\name{open.dir}
\alias{open.dir}
\title{open.dir}
\usage{
\method{open}{dir}(dir = ".")
}
\arguments{
\item{dir}{PARAM_DESCRIPTION, Default: '.'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
