% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_dir.R
\name{open_dir}
\alias{open_dir}
\title{open_dir}
\usage{
open_dir(dir = ".")
}
\arguments{
\item{dir}{PARAM_DESCRIPTION, Default: '.'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
