% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_help.R
\name{quick_help}
\alias{quick_help}
\title{quick_help}
\usage{
quick_help(person, project, path = getOption("cpmtools.quick_help_path"))
}
\arguments{
\item{person}{PARAM_DESCRIPTION}

\item{project}{PARAM_DESCRIPTION}

\item{path}{PARAM_DESCRIPTION, Default: getOption("cpmtools.quick_help_path")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

}
\seealso{
\code{\link[icesTAF]{taf.skeleton}}
\code{\link[git2r]{init}},\code{\link[git2r]{add}},\code{\link[git2r]{commit}}
}
