% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceTAF2.R
\name{sourceTAF2}
\alias{sourceTAF2}
\title{sourceTAF2}
\usage{
sourceTAF2(script, quiet = FALSE)
}
\arguments{
\item{script}{PARAM_DESCRIPTION}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
