% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf_gitignore.R
\name{taf_gitignore}
\alias{taf_gitignore}
\title{taf_gitignore}
\usage{
taf_gitignore(path = ".", append = FALSE)
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: '.'}

\item{append}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
