% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf_roxy.R
\name{taf_roxy}
\alias{taf_roxy}
\title{create a blank data set file}
\usage{
taf_roxy(dataset)
}
\arguments{
\item{dataset}{the name of the dataset that will be created}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
description
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
