% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.taf.json.R
\name{write.taf.json}
\alias{write.taf.json}
\title{write.taf.json}
\usage{
write.taf.json(x, file = NULL, dir = NULL, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION, Default: NULL}

\item{dir}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
no return value
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[jsonlite]{read_json}}
}
