% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sam_upload.R
\name{write_sam_upload}
\alias{write_sam_upload}
\title{Write SAM assesseent to a TAF json file}
\usage{
write_sam_upload(fit, dir = NULL)
}
\arguments{
\item{fit}{a fitted object from a SAM model fit.}

\item{dir}{an optional directory name.}
}
\description{
Write a SAM assessment to a json file int the correct
format to be uploaded to the TAF assessment results
database
}
