% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_r_win.R
\name{update_r_win}
\alias{update_r_win}
\title{update_r_win}
\usage{
update_r_win(check.only = TRUE)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
